chore: Move functions out of system_test_definition and into helper_functions
